Make sure Rental exists before trying to update it
In btnBrowse_Click, if Rental is null, tell the user only which inputs were invalid.
In btnUpdate_Click, when all inputs are invalid, tell the user all inputs can't be invalid.
Otherwise, see if there's only one Rental based on inputs provided. If not tell the user only which inputs were invalid.
If there is only one Rental found, fill in inputs that are invalid. Then try to update.
In CheckAll, the message should have always been displayed when greater than max, not greater than or equal to max.
Use minimum value of date for SQL (1/1/1753) instead of minimum value of date for C#.
Message should have always been "greater than ", not "greater than or equal to ".